mesh dit

+++ mmfreelm/models/hgrn_bit/mesh_dit.py
@@ -89,7 +89,7 @@ class TimestepEmbedder(nn.Module):
 
     def forward(self, t):
         t_freq = self.timestep_embedding(t, self.frequency_embedding_size)
-        t_emb = self.mlp(t_freq)
+        t_emb = self.mlp(t_freq.to(self.mlp[0].weight.dtype))
         return t_emb
